build: removing typescript from vue
I am removing typescript from the project, as it seems to be less
useful in vue with python. We don't have a way of keeping type safety
between the python api and the client. Maybe in the future we can add
it back but the cons outweigh the pros right now.